ADD test methods to CacheModelQueryTest for testing different ways of using CacheModelQuery
REFACTOR QueryCacheAttributeTest to CacheModelQueryTest.
